+++ src/gamdam/__init__.py
@@ -6,12 +6,10 @@ of JSON Lines describing what to download and what metadata each file has,
 downloads them in parallel to a git-annex_ repository, attaches the metadata
 using git-annex's metadata facilities, and commits the results.
 
-This program was written as an experiment/proof-of-concept for a larger
-program.  It is my first time using trio_, so there may be some oddness in the
-code.
+This program was written as an experiment/proof-of-concept for a larger program
+and is now only minimally maintained.
 
 .. _git-annex: https://git-annex.branchable.com
-.. _trio: https://github.com/python-trio/trio
 
 Visit <https://github.com/jwodder/gamdam> for more information.
 """