+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.2.0 (in development)
 -----------------------
 - Update for changes to the `registerurl` command in git-annex 10.20220222
     - `gamdam` now requires git-annex v10.20220222 or higher to run
+- Switch from trio to anyio
 
 v0.1.0 (2021-10-24)
 -------------------